aplicando disposição de pasta para iniciar arquitetura hexagonal
