+++ src/test/controllerTest/contas.controller.spec.ts
@@ -1,26 +1,32 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { ContasController } from '../../controller/contas.controller';
+import { ContasController } from '../../contas/aplication/contas.controller';
 import { AppModule } from '../../app.module';
 import supertest from 'supertest';
 import { INestApplication } from '@nestjs/common';
 import { TipoConta } from '../../enum/conta.enum';
-import { ContasService } from '../../service/contas.service';
-import { UserService } from '../../service/user.service';
-import { ContasFactory } from '../../factory/contas.factory';
-import { ContasRepository } from '../../repository/contas.repository';
-import { UserRepository } from '../../repository/user.repository';
-import { TUser } from 'src/model/user.entity';
+import { ContasService } from '../../contas/aplication/contas.service';
+import { UserService } from '../../user/aplication/user.service';
+import { ContasFactory } from '../../contas/domain/factory/contas.factory';
+import { ContasRepository } from '../../contas/adapter/contas.repository';
+import { TUser } from '../../user/domain/model/user.entity';
+import { UserRepository } from '../../user/adapter/user.repository';
 
 describe('ContasController', () => {
   let controller: ContasController;
-  let app: INestApplication 
-  let cliente: TUser
+  let app: INestApplication;
+  let cliente: TUser;
 
-    beforeEach(async () => {
-      const module: TestingModule = await Test.createTestingModule({
-      imports:[AppModule],
-      providers:[ContasService, UserService, UserRepository, ContasFactory, ContasRepository],
-      controllers:[ContasController]
+  beforeEach(async () => {
+    const module: TestingModule = await Test.createTestingModule({
+      imports: [AppModule],
+      providers: [
+        ContasService,
+        UserService,
+        UserRepository,
+        ContasFactory,
+        ContasRepository,
+      ],
+      controllers: [ContasController],
     }).compile();
 
     controller = module.get<ContasController>(ContasController);
@@ -30,21 +36,19 @@ describe('ContasController', () => {
 
   it('should be defined', () => {
     expect(controller).toBeDefined();
-    expect(controller).toBeInstanceOf(ContasController)
+    expect(controller).toBeInstanceOf(ContasController);
   });
 
   test('Deveria criar uma conta ', () => {
-    const tipo = TipoConta
+    const tipo = TipoConta;
 
-    
     return supertest(app.getHttpServer())
       .post('/contas')
-      .send({ tipo, cliente})
+      .send({ tipo, cliente })
       .expect(201)
       .expect(({ body }) => {
         expect(body.tipo).toBe(tipo);
         expect(body.cliente).toBe(cliente);
       });
   });
-
 });
