+++ src/test/serviceTest/contas.service.spec.ts
@@ -1,12 +1,12 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { ContasService } from '../../service/contas.service';
-import { ContasFactory } from '../../factory/contas.factory';
-import { ContasRepository } from '../../repository/contas.repository';
+import { ContasService } from '../../contas/aplication/contas.service';
+import { ContasFactory } from '../../contas/domain/factory/contas.factory';
+import { ContasRepository } from '../../contas/adapter/contas.repository';
 import { TipoConta } from '../../enum/conta.enum';
-import { ContaCorrente } from '../../model/contaFeature/contaCorrente';
-import { ContaPoupanca } from '../../model/contaFeature/contaPoupanca';
+import { ContaCorrente } from '../../contas/domain/model/feature/contaCorrente';
+import { ContaPoupanca } from '../../contas/domain/model/feature/contaPoupanca';
 import { uuid } from 'uuidv4';
-import { ClienteFisico } from '../../model/user-feature/clienteFisico.model';
+import { ClienteFisico } from '../../user/domain/model/feature/clienteFisico.model';
 
 jest.mock('uuidv4');
