+++ src/user/domain/model/feature/clienteJuridico.model.ts
