+++ src/user/domain/model/feature/gerente.model.ts
